move logic from saddles to its own InternalSaddleNetwork object.
